machine addition by users
Now user can add a machine among the 3 types available, the machines are not automatically added anymore. Users can also delete the machines, but only from the frontend, they are still stored in the db. The synthetic data is generated from a hardcoded file, depending on what machine type the user chose to add.